Created Structure model. Represents a physical tower/etc.
Also created a script to query the ASR database for a given facility and find any structures within X meters of it, then create a Structure from this information (or add to one that already exists)

Detail and List views are in place for Structure, but need some tweaking still.

Also still a ways to go on verifying that I'm actually matching things up correctly (what should the proper radius be? I was using 100m here, but will probably switch to 10)