Introduced concept of BatchAudit/BatchAuditGroup
Before a Batch creation is attempted, a BatchAudit is first created.
This will persist regardless of whether the Batch is rolled back. A
BatchAudit belongs to a BatchAuditGroup, and so a BatchAuditGroup must
be created if one does not already exist for the current Batch.

The function of the BatchAuditGroup is to handle the case where there
are a series of Batch imports attempted, and thus a series of
BatchAudits created, but they all fail. Without the BatchAuditGroup,
none of these BatchAudits are connected in any way (if there were a
Batch, then you could examine its Batch.audits).

So, now everything is linked together no matter what! Also, a
BatchAuditGroup has a status, like BatchAudit. This is to avoid
needlessly expensive queries. It is updated every time a BatchAudit is
saved.

Re-importing buttons exist in Batch/BatchAudit/BatchAuditGroup pages,
and all appear to be functional.

import_original_excel_application has been removed, because it was made
redundant by _another_ huge change: ExcelCollectionImporter/ExcelImporter
now accept a batch/batches in addition to path/paths. That is, they can
operate on either files or batch objects in memory. This avoids the need
for management of intermediary files, which was the sole function of
import_original_excel_application