Enabled table exports
Made adjustments to FilterTableView. Now allows children to define an optional export_table_class. If this is defined, it is used instead of the 'standard' table_class in order to export data. Basically this allows for a different set of fields to be exported than viewed.

Also added the necessary buttons to the form helpers

NRQZ-146