Added a bunch of fields to NRQZ Analyzer Importer
Lots of minor tweaks there and other places to accommodate

New thing: RowData now tracks row-level errors. Never had these before, but we have them now. Only example so far: if a case was created but shouldn't have been, we have a row-level error. This happens in all of the technical data-focused data. The idea is that all cases should be created from application data. If that is missing, we have a problem