Fix bug where attachment creation from DAL was broken
Basically, the user would paste in a path wrapped in quotes, which I
have logic in the Attachment.save method to strip out. But since that
logic isn't called until save-time, DAL didn't detect any attachments
with that path in the DB already! So, it would try to create a new one,
but when the quotes were stripped it was no longer unique, causing the
error. Quotes are now stripped in AttachmentAutocompleteView itself, and
things work correctly